+++ TestTools/StreamChatTestTools/Mocks/StreamChat/WebSocketClient/WebSocketClient_Mock.swift
@@ -21,8 +21,6 @@ final class WebSocketClient_Mock: WebSocketClient {
     var disconnect_called: Bool { disconnect_calledCounter > 0 }
     var disconnect_completion: (() -> Void)?
 
-    var timeout_callCount = 0
-
 
     var mockedConnectionState: WebSocketConnectionState?
 
@@ -76,10 +74,6 @@ final class WebSocketClient_Mock: WebSocketClient {
         disconnect_calledCounter += 1
         disconnect_source = source
         disconnect_completion = completion
-    }
-
-    override func timeout() {
-        timeout_callCount += 1
     }
 
     var mockEventsBatcher: EventBatcher_Mock {